mod(CMake): Generate simulation core version.h
The version can be specified on the CMake command line using the
following variables:
- SIMCORE_VERSION_MAJOR
- SIMCORE_VERSION_MINOR
- SIMCORE_VERSION_PATCH
- SIMCORE_VERSION_TAG

If all of major, minor and patch are defined, they are used as openPASS
simulation core version. Otherwise the tag string will be used. If the
tag is also undefined, the pre-defined string constant
"undefined-version" will be used.