mod(World_OSI): Consistent MovingObject order in SensorView generation
Simulated MovingObjects were stored in a std::unordered_map with an
integer key. This lead to undefined order of objects inside of
generated SensorViews starting from the second invocation of a
simulation. This is caused by the unordered maps retaining some internal
state, even when being cleared.

This commit changes the implementation to use std::map for
MovingObjects.

Signed-off-by: Reinhard Biegel <reinhard.biegel@in-tech.com>