fix(pyOpenPASS): Use integer format in query for detected agents field
Behavior of pandas has changed from 1.1.x to 1.2.x.
Comparisions between different datatypes in DataFrame.query() seem to be
more strict now, causing the query for the Sensor_Latency test to fail.

This commit solves the problem for the Sensor_Latency test by using an
integer instead of string for the expected Agent ID.

Signed-off-by: Reinhard Biegel <reinhard.biegel@in-tech.com>