fix(pyOpenPASS): Correctly build path to config under test
When setting the random seed for the determinism test the initial value
shall be read from the configuration file (slaveConfig.xml). The path to
this file was built incorrectly, resulting in random seed been taken
from the 'Common' configuration instead.

This misbehavior lead to failing determinism tests when the random seed
of a configuration didn't match the value specified in the 'Common'
configuration by coincidence.

Signed-off-by: Reinhard Biegel <reinhard.biegel@in-tech.com>